requirements.txt
streamlit
rdkit-pypi
scikit-learn
pandas
numpy
